Added physics sync start

+++ JobSimulatorMultiplayer/MonoBehaviours/IDHolder.cs
@@ -6,6 +6,6 @@ namespace JobSimulatorMultiplayer.MonoBehaviours
     public class IDHolder : MonoBehaviour
     {
         public IDHolder(IntPtr ptr) : base(ptr) { }
-        public short ID;
+        public int ID;
     }
 }
